Add docker compose files

+++ scripts/deploy.sh
@@ -4,16 +4,21 @@ IFS='
 export $(egrep -v '^#' ./.env | xargs -0) # load env
 IFS=
 
-rm -r build
-rm -r frontend/src/contracts
+if $DEPLOY
+then
+    rm -r build
+    rm -r frontend/src/contracts
 
-truffle migrate --network $NETWORK # deploy main contracts
+    truffle migrate --network $NETWORK # deploy main contracts
 
-cd ethr-did-registry
+    cd ethr-did-registry
 
-truffle migrate --network $NETWORK # deploy EthereumDIDRegistry contract
+    truffle migrate --network $NETWORK # deploy EthereumDIDRegistry contract
 
-cd ..
+    cd ..
 
-node scripts/seed.js
-node scripts/copy-contracts.js
+    node scripts/seed.js
+    node scripts/copy-contracts.js
+fi
+
+node server.js
